feat(controller): add wildcard subdomain support
This alters the check for wildcard subdomains in the controller to not
explicitly disallow wildcard subdomains.  It adds some basic checks to
make sure the wildcard is at the start, and to make sure that there is
no clash with any apps that are not being worked on by the current user.
It also ensures that the wildcard domain can not obstruct the deis
platform domain, as nginx always prefers wildcard domains over regex
domains (which the controller and all apps use by default).

I've been running wildcard domains in deis for some time by manually
adding them into etcd, so I'm reasonably confident that the controller
validation is the only thing preventing this from working.

This fixes #1608.